The tcp_backlog meter actually doesn't actually bear any fruit
For reasons of Linux, and, related to Puma things, this SYN_RECV
counter isn't an effective tool for measuring anything meaningful.
The Linux kernel actually has two queues, and the queue that is
important to measure ... can't be. So, this SYN_RECV counter is
really measuring the number of client initiated connections that
haven't completed the 3-way TCP handshake (e.g. haven't ACK'd),
and that's a typical thing of a malicious actor, or very slow
internet connection (but something we'd unlikely catch at a 20
second polling).

There's an item in Puma (and possibly other ruby web servers,
called the "backlog", which represents the number of connections
waiting for a thread in the thread pool. This is likely a more
interesting measurement, but even more interesting is probably
the amount of time those connections sit idle waiting to be
picked up, which we can get from the Heroku router. Since we're
targetting Heroku with this, we'll simply dismiss this as being
useful and move on.