Giant refactor!!!!!!!!!
Remove any and all assumptions that things are request driven. It's
all periodic, all the time.

In addition, simplify the concept of a meter, into a panel which
contains instruments that can be `instrument!`'d. These `!`methods
modify passed in state and are essentially folded over to collect
the current values of the gauges and counters. The instruments
themselves emit the proper key, and in the case of counters that
we'd like to treat as special counters, they apply sampling. See
an explanation of that in doc.rb.

Fix up the Railtie such that we can actually configure this thing
a bit easier, and remove things we'll just plain not utilize, like
the ChangeInstrument and GaugeInstrument which weren't utilized
and confused things.